Stop using sequelize operator aliases, as advised for security reasons
Sequelize docs advise against operator aliases, as they are a vector
opening up the possibility of injecting an Object with string operators
to Sequelize.
More info here:
http://docs.sequelizejs.com/manual/tutorial/querying.html#operators

On top of that, there are plans for deprecating them.

This changeset removes the use of sequelize operator aliases from the
package 'extensions'.